Renamed Midi Event Classes
