+++ src/core/midi/GOMidiReceiverEventPattern.h
