+++ src/core/midi/GOMidiSenderEventPattern.h
